Coloc 2: Add an implementation for Li's scatter plot

+++ src-plugins/Colocalisation_Analysis/Algorithms/Histogram2D.java
@@ -212,7 +212,7 @@ public class Histogram2D<T extends RealType<T>> extends Algorithm {
 	 * @param ch1BinWidt The bin width for channel one
 	 * @return The x value of the data point location
 	 */
-	protected int getXValue(double ch1Val, int ch1BinWidth, double ch2Val, int ch2BinWidth) {
+	protected int getXValue(double ch1Val, double ch1BinWidth, double ch2Val, double ch2BinWidth) {
 		return (int)(ch1Val * ch1BinWidth);
 	}
 
@@ -222,7 +222,7 @@ public class Histogram2D<T extends RealType<T>> extends Algorithm {
 	 * @param ch2BinWidt The bin width for channel one
 	 * @return The x value of the data point location
 	 */
-	protected int getYValue(double ch1Val, int ch1BinWidth, double ch2Val, int ch2BinWidth) {
+	protected int getYValue(double ch1Val, double ch1BinWidth, double ch2Val, double ch2BinWidth) {
 		return (yBins - 1) - (int)(ch2Val * ch2BinWidth);
 	}
 }
